currentTime(String pattern) I Basklass
